Implementata funzione per scrivere commenti

+++ app/src/main/java/com/example/dibapp/module/Comment.java
@@ -2,15 +2,15 @@ package com.example.dibapp.module;
 
 public class Comment {
     public String id;
-    public int rating;
+    public float rating;
     public String descrption;
     public boolean privatecomment;
 
-    public Comment (String id, int rating, String descrption, boolean privatecomment){
+    public Comment (String id, float rating, String descrption, boolean privatecomment){
         this.id=id;
         this.rating=rating;
         this.descrption=descrption;
         this.privatecomment=privatecomment;
     }
-    
+
 }
